@@ -5,11 +5,6 @@ import javassist.CtClass;
 import javassist.CtMethod;
 import javassist.CtNewMethod;
 
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.util.HashMap;
-import java.util.Map;
-
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -26,7 +21,7 @@ public class MapperGeneratorTest {
     CtMethod fooImpl = CtNewMethod.make("public String foo() { return \"bar\"; }", impl);
     impl.addMethod(fooImpl);
 
-    SimpleInterface si = (SimpleInterface) impl.toClass().newInstance();
+    SimpleInterface si = (SimpleInterface) impl.toClass().getDeclaredConstructor().newInstance();
     assertEquals("bar", si.foo());
   }
   
@@ -38,57 +33,34 @@ public class MapperGeneratorTest {
     CtClass impl = pool.makeClass("org.dt.japper.Mapper_123");
     impl.addInterface(intf);
     
-    CtMethod mapImpl = CtNewMethod.make("public Object map(java.sql.ResultSet rs, org.dt.japper.RowProcessor rowProcessor) { return \"bar\"; }", impl);
+    CtMethod mapImpl = CtNewMethod.make("public Object map(org.dt.japper.JapperConfig config, java.sql.ResultSet rs, org.dt.japper.RowProcessor rowProcessor) { return \"bar\"; }", impl);
     impl.addMethod(mapImpl);
     
     @SuppressWarnings("unchecked")
-    Mapper<String> mapper = (Mapper<String>) impl.toClass().newInstance();
+    Mapper<String> mapper = (Mapper<String>) impl.toClass().getDeclaredConstructor().newInstance();
     
-    assertEquals("bar", mapper.map(null, null));
+    assertEquals("bar", mapper.map(null,null, null));
   }
   
   
   @Test
   public void genericInterfaceWithLookupTest() throws Exception {
-    Mapper<String> mapper = createMapper(String.class).newInstance();
-    assertEquals("bar", mapper.map(null, null));
+    Mapper<String> mapper = createMapper(String.class).getDeclaredConstructor().newInstance();
+    assertEquals("bar", mapper.map(null,null, null));
   }
   
 
-  @SuppressWarnings({"unchecked", "cast"})
-  private <T> Class<Mapper<T>> createMapper(Class<T> resultType) throws Exception {
+  @SuppressWarnings({"unchecked", "cast", "SameParameterValue"})
+  private <T> Class<Mapper<T>> createMapper(Class<T> ignoredResultType) throws Exception {
     ClassPool pool = new ClassPool(true);
     CtClass intf = pool.get("org.dt.japper.Mapper");
     CtClass impl = pool.makeClass("org.dt.japper.Mapper_234");
     impl.addInterface(intf);
     
-    CtMethod mapImpl = CtNewMethod.make("public Object map(java.sql.ResultSet rs, org.dt.japper.RowProcessor rowProcessor) { return \"bar\"; }", impl);
+    CtMethod mapImpl = CtNewMethod.make("public Object map(org.dt.japper.JapperConfig config, java.sql.ResultSet rs, org.dt.japper.RowProcessor rowProcessor) { return \"bar\"; }", impl);
     impl.addMethod(mapImpl);
     
     return (Class<Mapper<T>>) impl.toClass();
-  }
-  
-  
-  private static class ByteClassLoader extends URLClassLoader {
-
-    private final Map<String, byte[]> extraClassDefs = new HashMap<String, byte[]>();
-    
-    public ByteClassLoader(URL[] urls, ClassLoader parent) {
-      super(urls, parent);
-    }
-
-    public void addClass(String name, byte[] bytecode) {
-      extraClassDefs.put(name, bytecode);
-    }
-    
-    @Override
-    protected Class<?> findClass(String name) throws ClassNotFoundException {
-      byte[] bytecode = this.extraClassDefs.remove(name);
-      if (bytecode != null) {
-        return defineClass(name, bytecode, 0, bytecode.length);
-      }
-      return super.findClass(name);
-    }
   }
   
 }
