@@ -10,23 +10,13 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-/**
- * Unit test for the BLOB writing.
- * <p>
- * Incidentally tests the reading of BLOB fields in order to actually assert that
- * the writing succeeded.
- *
- * @author David Sykes
- *
- */
-public class WriteBlobTest {
-
+public class ReadBlobTest {
   private static final String SQL_INSERT_ATTACHMENT =
           " INSERT INTO attachment(id, mime_type, attachment) VALUES(:ID, :MIME_TYPE, :ATTACHMENT)"
           ;
 
   private static final String SQL_ATTACHMENT =
-            " select id, mime_type, attachment "
+          " select id, mime_type, attachment "
           + "   from attachment "
           + "  where id = :ID"
           ;
@@ -45,24 +35,39 @@ public class WriteBlobTest {
   }
 
   @Test
-  public void writeBlobTest() throws Exception {
+  void check_blob_limit_is_respected() throws Exception {
     Connection conn = testData.connect();
 
-    byte[] blob = "MAGIC".getBytes(StandardCharsets.UTF_8);
+    insertTestAttachment(conn);
+
+    Attachment attachment = Japper.queryOne(conn, Attachment.class, SQL_ATTACHMENT
+            , "ID", 13
+    );
+    assertNotNull(attachment);
+
+    // Lower the limit below the length of our actual blob
+    // via the JapperConfig object
+    try {
+      Japper.queryOne(JapperConfig.maxBlob(25)
+              , conn, Attachment.class, SQL_ATTACHMENT
+              , "ID", 13
+      );
+      fail("Loaded BLOB larger than limit!");
+    }
+    catch (IllegalArgumentException iaEx) {
+      // If we get here, all is well!
+    }
+  }
+
+  private void insertTestAttachment(Connection conn) {
+    byte[] blob = "MAGIC MAGIC MAGIC 30+ bytes long".getBytes(StandardCharsets.UTF_8);
 
     int rowsAffected = Japper.execute(conn, SQL_INSERT_ATTACHMENT
             , "ID", 13
             , "MIME_TYPE", "text/plain"
             , "ATTACHMENT", blob
-            );
+    );
     assertEquals(1, rowsAffected);
 
-    Attachment a = Japper.queryOne(conn, Attachment.class, SQL_ATTACHMENT, "ID", 13);
-    assertNotNull(a);
-
-    String s = new String(a.getAttachment(), StandardCharsets.UTF_8);
-    assertEquals("MAGIC", s);
-
-    conn.close();
   }
 }
